добавил кусок админ панели

+++ src/main/java/Ivan/Vitsin/springsecurity/controller/AdminController.java
@@ -67,7 +67,7 @@ public class AdminController {
     }
 
     @DeleteMapping("/user-delete/{id}")
-    public String deleteUser(@PathVariable("id") int id) {
+    public String deleteUser(@PathVariable("id") Long id) {
         userService.deleteById(id);
         return  "redirect:/admin";
     }
