+++ src/main/java/Ivan/Vitsin/springsecurity/service/UserService.java
@@ -1,6 +1,5 @@
 package Ivan.Vitsin.springsecurity.service;
 
-import org.springframework.security.core.userdetails.UserDetailsService;
 import Ivan.Vitsin.springsecurity.model.User;
 
 import java.util.List;
@@ -8,13 +7,13 @@ import java.util.List;
 public interface UserService {
 
 
-    User findById(int id);
+    User findById(Long id);
 
     List<User> findAll();
 
     void saveUser(User user);
 
-    void deleteById(int id);
+    void deleteById(Long id);
 
     User findUserByEmail(String email);
